Switch away from GLib
Although GLib was helpful during early development, since it saved
us from writing custom data structures, there was an issue: GLib
follows an abort-on-failure memory allocation strategy, which isn't
really appropriate for potential uses of Ruby417 (i.e. web apps).

Still using GLib testing, however.

Implemented a dynamic array (DArray) that pretty much mirrors the
GLib GPtrArray API. Sorting is slower, though not critically.

Also reorganized source files.

+++ test/ext/run_suite.sh
@@ -5,22 +5,26 @@
 #
 # ext/
 #   ruby417/
+#     darray/
+#       darray.h
 #     rectangles/
-#       ...
+#       rectangles.c
 #     ...
-#   test/
+# test/
+#   ext/
 #     fixtures/
 #     run_suite.sh
 #     ...
+#
 
-function error() {
+function error {
   echo "$(basename $0): $1"
   exit -1
 }
 
-base_dir="$(dirname $0)/.."
-source_dir="$base_dir/ruby417"
-test_dir="$base_dir/test"
+base_dir="$(dirname $0)/../.."
+source_dir="$base_dir/ext/ruby417"
+test_dir="$base_dir/test/ext"
 out_file="test_$(date +%s).out"
 
 if [[ ! -e $source_dir/rectangles/rectangles.h ]]; then
@@ -36,14 +40,15 @@ if [[ ! -d $test_dir/fixtures ]]; then
 fi
 
 echo "Recompiling suite..."
-gcc $test_dir/tests.c $test_dir/utilities.c         \
-  `pkg-config --libs --cflags glib-2.0` -lm         \
-  -DNO_RUBY                                         \
+gcc $test_dir/tests.c $test_dir/utilities.c                                     \
+  `pkg-config --libs --cflags glib-2.0` -lm                                     \
+  -DNO_RUBY -DEXT_INCLUDE_DIR="\"../../ext/ruby417/rectangles/rectangles.c\""   \
   -o $test_dir/$out_file
 
-echo "Running tests..."
-(cd $test_dir; ./$out_file $@)
+if [[ -e $test_dir/$out_file ]]; then
+  echo "Running tests..."
+  (cd $test_dir; ./$out_file $@)
 
-rm $test_dir/$out_file
-
-echo "Done."
+  rm $test_dir/$out_file
+  echo "Done."
+fi
